% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_counts.R
\name{get_site_counts}
\alias{get_site_counts}
\title{Get Site Counts Matrix (Internal)}
\usage{
get_site_counts(partitions, site_index, alphabet_size)
}
\arguments{
\item{partitions}{List of data frames.}

\item{site_index}{Integer.}

\item{alphabet_size}{Integer.}
}
\value{
Numeric matrix (rows=AA, cols=Time).
}
\description{
Extract and tabulate amino acid counts for a specific site across all partitions.
}
\keyword{#}
\keyword{#'}
\keyword{'A'}
\keyword{'A's}
\keyword{'R's}
\keyword{(1)}
\keyword{(2)}
\keyword{(A)}
\keyword{(R)}
\keyword{(code}
\keyword{1}
\keyword{1)}
\keyword{1))}
\keyword{1,}
\keyword{1.}
\keyword{1:}
\keyword{1]}
\keyword{2}
\keyword{2))}
\keyword{2,}
\keyword{2.}
\keyword{25)}
\keyword{2:}
\keyword{2]}
\keyword{=}
\keyword{@examples}
\keyword{Create}
\keyword{Expect}
\keyword{Get}
\keyword{Note:}
\keyword{Partition}
\keyword{Row}
\keyword{Since}
\keyword{Site}
\keyword{ViralEntroR:::get_site_counts}
\keyword{[0,}
\keyword{[3,}
\keyword{])}
\keyword{alphabet_size}
\keyword{and}
\keyword{be}
\keyword{c(1,}
\keyword{call}
\keyword{counts}
\keyword{data.frame(s1}
\keyword{directly}
\keyword{dummy}
\keyword{for}
\keyword{get_site_counts(parts,}
\keyword{has}
\keyword{if}
\keyword{installed.}
\keyword{internal}
\keyword{internal,}
\keyword{is}
\keyword{it}
\keyword{list(p1,}
\keyword{loaded,}
\keyword{one}
\keyword{or}
\keyword{p1}
\keyword{p2}
\keyword{p2)}
\keyword{partitions}
\keyword{parts}
\keyword{print(counts[1:5,}
\keyword{site_index}
\keyword{this}
\keyword{three}
\keyword{to}
\keyword{two}
\keyword{use}
\keyword{we}
